+++ tests/commands/test_init.py
@@ -6,7 +6,7 @@ import tempfile
 from cleo import CommandTester
 
 
-def test_default(app, mocker, tmp_dir):
+def test_basic_interactive(app, mocker, tmp_dir):
     getcwd = mocker.patch('os.getcwd')
     getcwd.return_value = tmp_dir
 
@@ -25,19 +25,40 @@ def test_default(app, mocker, tmp_dir):
     tester.execute([('command', command.name)])
 
     output = tester.get_display()
-    expected = """
-Generated file
-
-[package]
-name = "my-package"
-version = "0.1.0"
-description = "This is a description"
-license = "MIT"
-
-
-
-Do you confirm generation [yes]? 
-"""
+    expected = command.template('poetry.toml').render(
+        name='my-package',
+        version='0.1.0',
+        description='This is a description',
+        authors=[],
+        license='MIT'
+    )
 
     assert expected in output
     assert os.path.exists(os.path.join(tmp_dir, 'poetry.toml'))
+
+
+def test_default_template(app, mocker, tmp_dir):
+    getcwd = mocker.patch('os.getcwd')
+    getcwd.return_value = tmp_dir
+
+    command = app.find('init')
+    tester = CommandTester(command)
+    tester.execute([('command', command.name), ('template', 'default')])
+
+    output = tester.get_display()
+    content = command.template('poetry.toml').render()
+    expected = """
+                                        
+  Welcome to the Poet config generator  
+                                        
+
+
+Using default template to create your poetry.toml config.
+
+"""
+
+    assert expected == output
+    assert os.path.exists(os.path.join(tmp_dir, 'poetry.toml'))
+
+    with open(os.path.join(tmp_dir, 'poetry.toml')) as f:
+        assert content == f.read()